test(webapp): split triggerTask engine test into per-concern files
The engine triggerTask suite was one 2447-line file with 23 containerTest
cases, each spinning its own Postgres + Redis. vitest shards by whole file,
so all 23 container setups landed on a single shard and dominated its
wall-clock. Local timings under-count this (containers start fast when warm),
so the duration-sharding sequencer treated the file as light and stacked it.

Split into four per-concern files sharing a triggerTaskTestHelpers module
(the vi.mock calls stay per-file, since they are hoisted):

- triggerTask.test.ts               trigger + idempotency + queue resolution
- triggerTask.debounce-mollifier    retries, debounce validation, mollifier
- triggerTask.metadataCache         DefaultQueueManager task metadata cache
- triggerTask.residency             child run residency inheritance

All 23 cases are preserved. Split the file's test-timings entry across the
four new files so sharding stays balanced.
